+++ backend/src/main/java/com/woowacourse/ternoko/comment/application/CommentService.java
@@ -9,11 +9,11 @@ import com.woowacourse.ternoko.comment.dto.CommentResponse;
 import com.woowacourse.ternoko.comment.dto.CommentsResponse;
 import com.woowacourse.ternoko.comment.exception.CommentNotFoundException;
 import com.woowacourse.ternoko.comment.repository.CommentRepository;
-import com.woowacourse.ternoko.domain.MemberType;
 import com.woowacourse.ternoko.domain.member.Member;
 import com.woowacourse.ternoko.interview.domain.Interview;
 import com.woowacourse.ternoko.interview.domain.InterviewRepository;
 import com.woowacourse.ternoko.interview.exception.InterviewNotFoundException;
+import com.woowacourse.ternoko.login.domain.MemberType;
 import com.woowacourse.ternoko.repository.CrewRepository;
 import java.util.ArrayList;
 import java.util.List;
