+++ backend/src/main/java/com/woowacourse/ternoko/login/exception/CoachNotAllowedException.java
@@ -1,7 +1,7 @@
 package com.woowacourse.ternoko.login.exception;
 
-import com.woowacourse.ternoko.common.exception.BadRequestException;
 import com.woowacourse.ternoko.common.exception.ExceptionType;
+import com.woowacourse.ternoko.common.exception.form.BadRequestException;
 
 public class CoachNotAllowedException extends BadRequestException {
     public CoachNotAllowedException(ExceptionType type) {
